requirements.txt
fastapi
uvicorn[standard]
openai>=1.0.0
pydantic
python-dotenv
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx
pandas
openpyxl>=3.0.0
xlrd>=2.0.0
numpy
python-multipart
azure-identity>=1.14.0
pyarrow
rapidfuzz
pytest
pytest-mock
pytest-asyncio
requests
aiohttp>=3.0.0
redis>=4.0.0
# Azure dependencies
azure-storage-blob>=12.0.0
azure-identity>=1.0.0
azure-keyvault-secrets>=4.0.0
# Microsoft Authentication Library
msal>=1.28.0
# SQL Server dependencies
pyodbc>=4.0.0
sqlalchemy>=2.0.0
# Cryptography for RSA encryption
cryptography>=41.0.0